% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHIstatistic.R
\name{BHI}
\alias{BHI}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
BHI(X)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \frac{1}{n {n\choose{2}}} \sum_{\mathcal{I}_{2}}
        \sum_{i_{3}=1}^n \left( \frac12 I\{|X_{i_1}| <
        |X_{i_{3}}|\} + \frac12 I\{|X_{i_2}| <
        |X_{i_{3}}|\} - I\{|X_{(2),X_{i_1},X_{i_2}}| < |X_{i_{3}}|\}\right) }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
BHI(X)
}
