% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.symmetry.R
\name{detect.symmetry}
\alias{detect.symmetry}
\title{detect.symmetry}
\usage{
detect.symmetry(A, sym.plane = "xz", non.sym = NULL, plot = TRUE)
}
\description{
This function detects paired landmarks in a dataset with object symmetry, similarly to MorphoJ (Klingenberg, 2011).
}
\details{
This function returns the paired landmarks.
}
\examples{
    detect.symmetry(A, sym.plane = "xz", non.sym = NULL, plot = TRUE)
  
}
\references{
Klingenberg, C.P. (2011) MorphoJ: an integrated software package for geometric morphometrics. Molecular ecology resources. 11: 353-357
}
\author{
Marta Vidal-Garcia, \email{marta.vidalga@gmail.com}
}
