% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGstatistic.R
\name{HG}
\alias{HG}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
HG(X, k, t, H = TRUE)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}

\item{k}{the value of parameter 'k' used in the formula}

\item{t}{the value with which to compare in indicator}

\item{H}{whether to calculate H or G}
}
\value{
The value of the statistics given by the formula:
        \deqn{ \frac{1}{{n\choose{2k+1}}} \sum_{\mathcal{I}_{2k+1}}
        I\{X_{(k+1),X_{i_1},\ldots,X_{i_{2k+1}}} < t }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
HG(X, 2)
}
