% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2statistic.R
\name{I2}
\alias{I2}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
I2(X)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \frac{1}{n^4} \sum\limits_{i,j,a,b=1}^n I\{|X_i - X_j| <
        X_a+X_b\}- I\{|X_i + X_j| < X_a+X_b\}}
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
I2(X)
}
