% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2HUAstatistic.R
\name{I2HUA}
\alias{I2HUA}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
I2HUA(X)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \frac{1}{n^2{n\choose2}} \sum_{1\leq i < j \leq n}\sum_{a,b=1}^n
        I\{|X_i - X_j| < X_a+X_b\}- I\{|X_i + X_j| < |X_a+X_b|\} }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
I2HUA(X)
}
