% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2USstatistic.R
\name{I2US}
\alias{I2US}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
I2US(X)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \frac{1}{{n\choose4}}{\sum_{1\leq i<j<a<b\leq n}
        I\{|X_i - X_j| < X_a+X_b\}- I\{|X_i + X_j| < X_a+X_b\}} }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
I2US(X)
}
