% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K1statistic.R
\name{K1}
\alias{K1}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
K1(X, k)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}

\item{k}{the value of parameter 'k' used in the formula}
}
\value{
The value of the test statistic given by the formula:
        \eqn{ \sup\limits_{t>0}\left|\frac{1}{\binom{n}{2k}}
        \sum\limits_{\mathcal{I}_{2k}} I\{|X_{(k),X_{i_1},\ldots,X_{i_{2k}}}|
        < t\}- I\{|X_{(k+1),X_{i_1},\ldots,X_{i_{2k}}}| < t \right|\} }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
K1(X, 2)
}
