% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2statistic.R
\name{K2}
\alias{K2}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
K2(X)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \sup\limits_{t>0}\frac{1}{n^2} \left| \sum\limits_{i,j=1}^n
        I\{|X_i - X_j| < t\}- I\{|X_i + X_j| < t\}\right| }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
K2(X)
}
