% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2Ustatistic.R
\name{K2U}
\alias{K2U}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
K2U(X)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \sup\limits_{t>0}\frac{1}{n\choose2} \left|
        \sum\limits_{1\leq i < j \leq n }
        I\{|X_i - X_j| < t\}- I\{|X_i + X_j| < t\}\right| }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
K2U(X)
}
