% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSWstatistics.R
\name{KS}
\alias{KS}
\title{Calculate Kolmogorov Smirnov test statistic (see 'Value' for formula)}
\usage{
KS(X)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \sup_t\left|F_n(t)-(1-F_n(-t))\right| }
}
\description{
Calculate Kolmogorov Smirnov test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
KS(X)
}
