% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstatistic.R
\name{M}
\alias{M}
\title{Calculate Mira test statistic (see 'Value' for formula)}
\usage{
M(X)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}
}
\value{
The value of the test statistic given by the formula:
< to be added >
}
\description{
Calculate Mira test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
M(X)
}
