% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIstatistic.R
\name{MI}
\alias{MI}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
MI(X, k)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}

\item{k}{the value of parameter 'k' used in the formula}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \frac{1}{n {n\choose{2k+1}}} \sum_{\mathcal{I}_{2k}}
        \sum_{i_{2k+1}=1}^n I\{-(X_{(k+1),X_{i_1},\ldots,X_{i_{2k}}}| <
        X_{i_{2k+1}}\}- I\{X_{(k+1),X_{i_1},\ldots,X_{i_{2k}}} < X_{i_{2k+1}}\} }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
MI(X, 2)
}
