% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAKstatistic.R
\name{NAK}
\alias{NAK}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
NAK(X, k)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}

\item{k}{the value of parameter 'k' used in the formula}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \sup_{t>0}\left|\frac{1}{{n\choose{k}}}
        \sum_{\mathcal{I}_{k}} I\{|X_{(1),X_{i_1},\ldots,X_{i_{k}}}|
        < t\}- I\{|X_{(k),X_{i_1},\ldots,X_{i_{k}}}| < t \}\right| }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
NAK(X, 2)
}
