% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvalues.R
\name{parTvalues}
\alias{parTvalues}
\title{Simulate the distribution of a test statistic in parallel}
\usage{
parTvalues(N, n, dist = list(), TS = list(), freecores = 0,
  clust = NULL)
}
\arguments{
\item{N}{the number of simulations to do}

\item{n}{the sample size for each simulation}

\item{dist}{a list which specifies the null distribution (see details)}

\item{TS}{a list which specifies the test statistic to use (see details)}

\item{freecores}{how many cores to leave unused (0 for maximum use of cpu)}

\item{clust}{a cluster to use for parallel}
}
\value{
A vector of size N, each element being the value of the statistic
        TS on simulated samples of size n.
}
\description{
This is just a parallel version of the \code{\link{Tvalues}} function, all
arguments apply for this function. See \code{\link{Tvalues}}.
}
\examples{
parTvalues(1000, 50, list(name='norm'), list(name='I1', k=2))
parTvalues(1000, 50, list(name='unif', min=-1, max=1), list(name='I2'))
parTvalues(1000, 50, list(name='logis', loc=0.5), list(name='K1', k=2))
parTvalues(1000, 50, list(name='exp'), list(name='K2'))
}
