% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sym.R
\name{plot.symmetry}
\alias{plot.symmetry}
\title{plot.symmetry}
\usage{
plot.symmetry(A, sym.pairs, sym.plane = NULL)
}
\description{
This function plots paired landmarks in a dataset with object symmetry, similarly to MorphoJ (Klingenberg, 2011).
}
\examples{
    left <- c(58, 15, 19, 21, 23, 25, 27, 31, 33, 35, 37, 43, 39, 45, 47, 49, 41, 29, 17, 6, 4, 8, 10, 12, 63, 62, 61, 60, 59, 51)
    right <- c(57, 14, 18, 20, 22, 24, 26, 30, 32, 34, 36, 42, 38, 44, 46, 48, 40, 28, 16, 7, 5, 9, 11, 13, 56, 55, 54, 53, 52, 50)
    sym.pairs <- cbind(left, right)
    plot.sym(A, sym.pairs)
    plot.sym(A, sym.pairs, sym.plane = "yz")
    plot.sym(A, sym.pairs, sym.plane = "xy")

}
\references{
Klingenberg, C.P. (2011) MorphoJ: an integrated software package for geometric morphometrics. Molecular ecology resources. 11: 353-357
}
\author{
Marta Vidal-Garcia, \email{marta.vidalga@gmail.com}
}
