% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rmixnorm}
\alias{rmixnorm}
\title{Mixture of 2 normal distributions}
\usage{
rmixnorm(n, mean1 = 0, sd1 = 1, mean2 = 0, sd2 = 1, p = 0.5)
}
\arguments{
\item{n}{number of observations}

\item{mean1}{mean of the first normal}

\item{sd1}{standard deviation of the first normal}

\item{mean2}{mean of the second normal}

\item{sd2}{standard deviation of the second normal}

\item{p}{probability of the first normal}
}
\value{
Vector of random numbers from the specified mixture of normals.
}
\description{
Generates random numbers from a mixture of 2 normal distributions
}
