% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rsl}
\alias{rsl}
\title{Azzalini skew logistic distribution}
\usage{
rsl(n = 1, xi = 0, omega = 1, alpha = 0, dp = NULL)
}
\arguments{
\item{n}{sample size.}

\item{xi}{ vector of location parameters.}

\item{omega}{vector of (positive) scale parameters.}

\item{alpha}{vector of slant parameters.}

\item{dp}{a vector of length 3 whose elements represent the parameters 
   described above. If \code{dp} is specified, the individual parameters 
   cannot be set.}
}
\value{
Vector of random numbers from Azzalini skew logistic distribution.
}
\description{
Generates random numbers from the skew logistic distribution
}
