% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSWstatistics.R
\name{SGN}
\alias{SGN}
\title{Calculate Signed test statistic (see 'Value' for formula)}
\usage{
SGN(X, mu = 0)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}

\item{mu}{the estimate of the location parameter}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \frac1n \sum_{i=1}^nI\{X_i - \mu > 0\} - \frac12}
}
\description{
Calculate Signed test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
SGN(X)
}
