% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\docType{package}
\name{symmetry}
\alias{symmetry}
\alias{symmetry-package}
\title{symmetry: A package which implements tests for symmetry`}
\description{
symmetry: A package which implements tests for symmetry`
}
