% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2statistic.R
\name{T2}
\alias{T2}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
T2(X, k)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}

\item{k}{the tuning parameter for the Laplace transform}
}
\value{
The value of the test statistic given by the formula:
        <to be added>
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
T2(X, 5)
}
