% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{test_power}
\alias{test_power}
\title{Calculate the power of a test}
\usage{
test_power(t0, t1, alpha = 0.05, two_sided = FALSE)
}
\arguments{
\item{t0}{the vector of null T values}

\item{t1}{the vector of alternative T values}

\item{alpha}{the significance level}

\item{two_sided}{indicator wheether to use two sided critical region}
}
\description{
This function calculates the power of a test given the null and alternative
T values and the significance level.
}
