% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_statistics_docs.R, R/B1statistic.R,
%   R/BHIstatistic.R, R/BHKstatistic.R, R/CHstatistic.R, R/CMstatistic.R,
%   R/K2Ustatistic.R, R/K2statistic.R, R/KSWstatistics.R, R/L2statistic.R,
%   R/MGGstatistic.R, R/MIstatistic.R, R/MKstatistic.R, R/MOIstatistic.R,
%   R/MOKstatistic.R, R/Mstatistic.R, R/NAIstatistic.R, R/NAKstatistic.R,
%   R/S1statistic.R, R/S2statistic.R, R/T1statistic.R, R/T2statistic.R
\name{TestStatistics}
\alias{TestStatistics}
\alias{B1}
\alias{BHI}
\alias{BHK}
\alias{CH}
\alias{CM}
\alias{K2U}
\alias{K2}
\alias{KS}
\alias{SGN}
\alias{WCX}
\alias{L2}
\alias{MGG}
\alias{MI}
\alias{MK}
\alias{MOI}
\alias{MOK}
\alias{M}
\alias{NAI}
\alias{NAK}
\alias{S1}
\alias{S2}
\alias{T1}
\alias{T2}
\title{Available test statistics for symmetry tests}
\usage{
B1(X)

BHI(X)

BHK(X)

CH(X)

CM(X)

K2U(X)

K2(X)

KS(X)

SGN(X)

WCX(X)

L2(X, k)

MGG(X)

MI(X, k)

MK(X, k)

MOI(X, k)

MOK(X, k)

M(X)

NAI(X, k)

NAK(X, k)

S1(X, k)

S2(X, k)

T1(X, k)

T2(X, k)
}
\arguments{
\item{X}{the numeric vector for which to calculate the test statistic}

\item{k}{the 'k' parameter in the formula (if applicable)}
}
\value{
The value of the test statistic.
}
\description{
The list of implemented test statistics and their functions
}
\details{
Below is a list of the implemented test statistics in the package. Each
statistic is listed by it's name, a code string (e.g. 'B1', CM','MOI') and
the formula of the statistic which is evaluated. The code string is used as
an argument to the \link{symmetry_test} function. Some statistics depend on a
parameter 'k' which can be seen from the formulas and is also passed as an
argument.

Each statistic is implemented as a function with the same name as the code
string, so the name of the function is passed as the argument "stat" to the
\link{symmetry_test} function
}
\section{Functions}{
\itemize{
\item \code{B1}: The \eqn{\sqrt{b_1}}  test statistic \insertCite{Cabilio1996}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{BHI}: The Cabilio–Masaro test statistic \insertCite{Cabilio1996}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{BHK}: The Cabilio–Masaro test statistic \insertCite{Cabilio1996}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{CH}: The Cabilio–Masaro test statistic \insertCite{Cabilio1996}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{CM}: The Cabilio–Masaro test statistic \insertCite{Cabilio1996}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{K2U}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{K2}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{KS}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{SGN}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{WCX}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{L2}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{MGG}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{MI}: The Mira test statistic \insertCite{Mira1999}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{MK}: The Mira test statistic \insertCite{Mira1999}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{MOI}: The Mira test statistic \insertCite{Mira1999}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{MOK}: The Mira test statistic \insertCite{Mira1999}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{M}: The Mira test statistic \insertCite{Mira1999}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{NAI}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{NAK}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{S1}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{S2}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{T1}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}

\item \code{T2}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

\deqn{
 \sum_{i=1}^ni^2
}
}}

\references{
\insertAllCited{}
}
