% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvalues.R
\name{Tvalues}
\alias{Tvalues}
\title{Simulate the distribution of a test statistic}
\usage{
Tvalues(N, n, dist = list(), TS = list())
}
\arguments{
\item{N}{the number of simulations to do}

\item{n}{the sample size for each simulation}

\item{dist}{a list which specifies the null distribution (see details)}

\item{TS}{a list which specifies the test statistic to use (see details)}
}
\value{
A vector of size N, each element being the value of the statistic
        TS on simulated samples of size n.
}
\description{
Simulates the distribution of the specified test statistic under the given
null distribution.
}
\details{
The dist argument is a list which must contain a field called "name" which
determines which distribution to use (e.g. "norm", "unif", "exp", etc.) and,
if needed, the parameters for the distribution. The name must be such that
the function "r"+name exists ("rnorm", "runif", "rexp", etc). Further
parameters are passed to that function.

The TS argument is a list which must contain a field called "name" which
specifies which test statistic function to use for each sample. The name can
be "I1", "K1", "I2", "K2" for statistics implemented by us, or any other
statistic for which an R function exists (e.g. "mean", "var", etc.).
}
\examples{
Tvalues(1000, 50, list(name='norm'), list(name='I1', k=2))
Tvalues(1000, 50, list(name='unif', min=-1, max=1), list(name='I2'))
Tvalues(1000, 50, list(name='logis', loc=0.5, sca=1), list(name='K1', k=2))
Tvalues(1000, 50, list(name='exp'), list(name='K2'))
}
