% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSWstatistics.R
\name{WCX}
\alias{WCX}
\title{Calculate Wilcoxon test statistic (see 'Value' for formula)}
\usage{
WCX(X, mu = 0)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}

\item{mu}{the estimate of the location parameter}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \frac{1}{{n\choose2}} \sum_{1\leq i<j\leq n}
        I\{X_i+X_j - 2\mu> 0\} - \frac12 }
}
\description{
Calculate Wilcoxon test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
WCX(X)
}
