% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1statistic.R
\name{B1}
\alias{B1}
\title{Calculate \eqn{\sqrt{b_1}} test statistic (see 'Value' for formula)}
\usage{
B1(X, mu = 0)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}

\item{mu}{the estimate of the location parameter}
}
\value{
The value of the test statistic given by the formula:
< to be added >
}
\description{
Calculate \eqn{\sqrt{b_1}} test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
B1(X, 2)
}
