% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHKstatistic.R
\name{BHK}
\alias{BHK}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
BHK(X)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
BHK(X)
}
