% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMstatistic.R
\name{CM}
\alias{CM}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
CM(X, mu = 0)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}

\item{mu}{the estimate of the location parameter}
}
\value{
The value of the test statistic given by the formula:
        \deqn{ \frac{1}{n {n\choose{2k}}} \sum_{\mathcal{I}_{2k}}
        \sum_{i_{2k+1}=1}^n I\{|X_{(k),X_{i_1},\ldots,X_{i_{2k}}} - \mu| <
        |X_{i_{2k+1}} - \mu|\}- I\{|X_{(k+1),X_{i_1},\ldots,X_{i_{2k}}} -
        \mu| < |X_{i_{2k+1}} - \mu|\} }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
CM(X, 2)
}
