% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMstatistic.R
\name{CM}
\alias{CM}
\title{Calculate Cabilio–Masaro test statistic (see 'Value' for formula)}
\usage{
CM(X)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}
}
\value{
The value of the test statistic given by the formula:
        <to be added>
}
\description{
Calculate Cabilio–Masaro test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
CM(X)
}
