% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1statistic.R
\name{L1}
\alias{L1}
\title{Calculate _ test statistic (see 'Value' for formula)}
\usage{
L1(X, a)
}
\arguments{
\item{X}{the sample for which to calculate the statistic}

\item{a}{the tuning parameter for the Laplace transform}
}
\value{
The value of the test statistic given by the formula: \deqn{\begin{array}{rl}
  \frac{1}{n^4} \sum_{i,j,k,l=1}^n &\left(\frac{1}{a+|X_i - X_j|+|X_k -
  X_l|}-\frac{1}{a+|X_i - X_j|+|X_k + X_l|}\right. \\ &-\left.\frac{1}{a+|X_i + X_j|+|X_k -
  X_l|} +\frac{1}{a+|X_i + X_j|+|X_k + X_l|}\right)\end{array} }
}
\description{
Calculate _ test statistic (see 'Value' for formula)
}
\examples{
set.seed(1)
X <- rnorm(50)
L1(X, 5)
}
