% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\docType{package}
\name{symmetry}
\alias{symmetry}
\alias{symmetry-package}
\title{symmetry: A package which implements tests for symmetry of random samples,
linear models and generalized autoregressive conditional heteroskedasticity
(GARCH) models}
\description{
The package contains a large number of tests for symmetry (and their
bootstrap variants), which can be used to test the symmetry of random samples
or of model residuals. Currently, the supported models are linear models and
generalized autoregressive conditional heteroskedasticity (GARCH) models
(fitted with the 'fGarch' package). The tests are implemented using the
'Rcpp' package which ensures great performance.
}
\details{
To see the available tests, see \link{TestStatistics}

For documentation on how to perform the tests, see \link{symmetry_test}
}
