% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_statistics_docs.R, R/B1statistic.R,
%   R/BHIstatistic.R, R/BHKstatistic.R, R/CHstatistic.R, R/CMstatistic.R,
%   R/K2Ustatistic.R, R/K2statistic.R, R/KSWstatistics.R, R/L2statistic.R,
%   R/MGGstatistic.R, R/MIstatistic.R, R/MKstatistic.R, R/MOIstatistic.R,
%   R/MOKstatistic.R, R/Mstatistic.R, R/NAIstatistic.R, R/NAKstatistic.R,
%   R/S1statistic.R, R/S2statistic.R, R/T1statistic.R, R/T2statistic.R
\name{TestStatistics}
\alias{TestStatistics}
\alias{B1}
\alias{BHI}
\alias{BHK}
\alias{CH}
\alias{CM}
\alias{K2U}
\alias{K2}
\alias{KS}
\alias{SGN}
\alias{WCX}
\alias{L2}
\alias{MGG}
\alias{MI}
\alias{MK}
\alias{MOI}
\alias{MOK}
\alias{M}
\alias{NAI}
\alias{NAK}
\alias{S1}
\alias{S2}
\alias{T1}
\alias{T2}
\title{Available test statistics for symmetry tests}
\usage{
B1(X)

BHI(X)

BHK(X)

CH(X)

CM(X)

K2U(X)

K2(X)

KS(X)

SGN(X)

WCX(X)

L2(X, k)

MGG(X)

MI(X, k)

MK(X, k)

MOI(X, k)

MOK(X, k)

M(X)

NAI(X, k)

NAK(X, k)

S1(X, k)

S2(X, k)

T1(X, k)

T2(X, k)
}
\arguments{
\item{X}{the numeric vector for which to calculate the test statistic}

\item{k}{the 'k' parameter in the formula (if applicable)}
}
\value{
The value of the test statistic.
}
\description{
The list of implemented test statistics and their functions
}
\details{
Below is a list of the implemented test statistics in the package. Each
statistic is listed by it's name, a code string (e.g. 'B1', CM','MOI') and
the formula of the statistic which is evaluated. The code string is used as
an argument to the \link{symmetry_test} function. Some statistics depend on a
parameter 'k' which can be seen from the formulas and is also passed as an
argument.

Each statistic is implemented as a function with the same name as the code
string, so the name of the function is passed as the argument "stat" to the
\link{symmetry_test} function
}
\section{Functions}{
\itemize{
\item \code{B1}: The \eqn{\sqrt{b_1}}  test statistic \insertCite{Cabilio1996}{symmetry}

to be added

\item \code{BHI}: The Cabilio–Masaro test statistic \insertCite{Cabilio1996}{symmetry}

\deqn{ \frac{1}{n {n\choose{2}}} \sum_{\mathcal{I}_{2}} \sum_{i_{3}=1}^n \left( \frac12 I\{|X_{i_1}| < |X_{i_{3}}|\} + \frac12 I\{|X_{i_2}| < |X_{i_{3}}|\} - I\{|X_{(2),X_{i_1},X_{i_2}}| < |X_{i_{3}}|\}\right) }

\item \code{BHK}: The Cabilio–Masaro test statistic \insertCite{Cabilio1996}{symmetry}

\deqn{ \sup_{t>0}\left|\frac{1}{{n\choose{2}}} \sum_{\mathcal{I}_{2}}\left( \frac12 I\{|X_{i_1}| <t\} + \frac12 I\{|X_{i_2}| <t\} - I\{|X_{(2),X_{i_1},X_{i_2}}| < t\}\right)\right| }

\item \code{CH}: The Cabilio–Masaro test statistic \insertCite{Cabilio1996}{symmetry}

to be added

\item \code{CM}: The Cabilio–Masaro test statistic \insertCite{Cabilio1996}{symmetry}

to be added

\item \code{K2U}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

\deqn{ \sup_{t>0}\frac{1}{ {n\choose 2}} \left| \sum_{1\leq i < j \leq n } I\{|X_i - X_j| < t\}- I\{|X_i + X_j| < t\}\right| }

\item \code{K2}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

\deqn{ \sup_{t>0}\frac{1}{n^2} \left| \sum_{i,j=1}^n I\{|X_i - X_j| < t\}- I\{|X_i + X_j| < t\}\right| }

\item \code{KS}: Kolmogorov--Smirnov test statistic \insertCite{Miao2006}{symmetry}

\deqn{ \sup_t\left|F_n(t)-(1-F_n(-t))\right| }

\item \code{SGN}: The Sign test statistic \insertCite{Miao2006}{symmetry}

\deqn{ \frac1n \sum_{i=1}^nI\{X_i > 0\} - \frac12}

\item \code{WCX}: The Wilcoxon test statistic \insertCite{Miao2006}{symmetry}

\deqn{ \frac{1}{{n\choose2}} \sum_{1\leq i<j\leq n} I\{X_i+X_j> 0\} - \frac12 }

\item \code{L2}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

\deqn{\begin{array}{rl} \frac{1}{n^4} \sum_{i,j,m,l=1}^n &\left(\frac{1}{k+|X_i - X_j|+|X_m - X_l|}-\frac{1}{k+|X_i - X_j|+|X_m + X_l|}\right. \\ &-\left.\frac{1}{k+|X_i + X_j|+|X_m - X_l|} +\frac{1}{k+|X_i + X_j|+|X_m + X_l|}\right)\end{array} }

\item \code{MGG}: The Miao, Gel and Gastwirth test statistic \insertCite{Miao2006}{symmetry}

to be added

\item \code{MI}: The Mira test statistic \insertCite{Mira1999}{symmetry}

\deqn{ \frac{1}{n {n\choose{2k+1}}} \sum_{\mathcal{I}_{2k}} \sum_{i_{2k+1}=1}^n I\{-(X_{(k+1),X_{i_1},\ldots,X_{i_{2k}}}| < X_{i_{2k+1}}\}- I\{X_{(k+1),X_{i_1},\ldots,X_{i_{2k}}} < X_{i_{2k+1}}\} }

\item \code{MK}: The Mira test statistic \insertCite{Mira1999}{symmetry}

\deqn{ \sup_{t>0}\left|\frac{1}{{n\choose{2k}}} \sum_{\mathcal{I}_{2k}} I\{-(X_{(k+1),X_{i_1},\ldots,X_{i_{2k}}}) < t\}- I\{X_{(k+1),X_{i_1},\ldots,X_{i_{2k}}} < t \}\right| }

\item \code{MOI}: The Mira test statistic \insertCite{Mira1999}{symmetry}

\deqn{ \frac{1}{n {n\choose{2k}}} \sum_{\mathcal{I}_{2k}} \sum_{i_{2k+1}=1}^n I\{|X_{(k),X_{i_1},\ldots,X_{i_{2k}}}| < |X_{i_{2k+1}}|\}- I\{|X_{(k+1),X_{i_1},\ldots,X_{i_{2k}}}| < |X_{i_{2k+1}}|\} }

\item \code{MOK}: The Mira test statistic \insertCite{Mira1999}{symmetry}

\deqn{ \sup_{t>0}\left|\frac{1}{{n\choose{2k}}} \sum_{\mathcal{I}_{2k}} I\{|X_{(k),X_{i_1},\ldots,X_{i_{2k}}}| < t\}- I\{|X_{(k+1),X_{i_1},\ldots,X_{i_{2k}}}| < t \}\right| }

\item \code{M}: The Mira test statistic \insertCite{Mira1999}{symmetry}

to be added

\item \code{NAI}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

\deqn{ \frac{1}{n {n\choose{k}}} \sum_{\mathcal{I}_{k}} \sum_{i_{k+1}=1}^n I\{|X_{(1),X_{i_1},\ldots,X_{i_{k}}}| < |X_{i_{k+1}}|\}- I\{|X_{(k),X_{i_1},\ldots,X_{i_{k}}}| < |X_{i_{k+1}}|\} }

\item \code{NAK}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

\deqn{ \sup_{t>0}\left|\frac{1}{{n\choose{k}}} \sum_{\mathcal{I}_{k}} I\{|X_{(1),X_{i_1},\ldots,X_{i_{k}}}| < t\}- I\{|X_{(k),X_{i_1},\ldots,X_{i_{k}}}| < t \}\right| }

\item \code{S1}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

to be added

\item \code{S2}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

to be added

\item \code{T1}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

to be added

\item \code{T2}: Allison \eqn{T_2} statistic \insertCite{Klar2012}{symmetry}

to be added
}}

\references{
\insertAllCited{}
}
